requirements.txt
flask
mysql.connection